# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API Laravel",
        "description": "Documentación generada automáticamente con Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alfred/transfer-requests": {
            "post": {
                "tags": [
                    "Alfred"
                ],
                "summary": "Crear solicitud de transferencia",
                "description": "Registra una solicitud de transferencia entre sender y receiver por teléfono. Estado inicial: solicitada.",
                "operationId": "alfredCreateTransferRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sender_phone",
                                    "receiver_phone",
                                    "amount"
                                ],
                                "properties": {
                                    "sender_phone": {
                                        "description": "Teléfono del remitente",
                                        "type": "string",
                                        "example": "18093901572"
                                    },
                                    "receiver_phone": {
                                        "description": "Teléfono del destinatario",
                                        "type": "string",
                                        "example": "18294428902"
                                    },
                                    "amount": {
                                        "description": "Monto solicitado en USD",
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "currency": {
                                        "description": "Moneda (default USD)",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Transferencia familiar",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "transfer_request": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "uuid": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "sender_phone": {
                                                    "type": "string",
                                                    "example": "18093901572"
                                                },
                                                "receiver_phone": {
                                                    "type": "string",
                                                    "example": "18294428902"
                                                },
                                                "amount": {
                                                    "type": "string",
                                                    "example": "150.00"
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "solicitada",
                                                        "rechazada",
                                                        "aprobado",
                                                        "completada",
                                                        "cancelada"
                                                    ],
                                                    "example": "solicitada"
                                                },
                                                "sender_customer_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "nullable": true
                                                },
                                                "receiver_customer_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "nullable": true
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}